refactor(User Story Vote): extract User Story Vote into its own Controller
