+++ Sources/App/Models/UserStoryVote.swift
@@ -17,8 +17,16 @@ extension UserStory {
             self.points[participant] = points
         }
 
-        var sum: Int { points.reduce(0, { $0 + $1.value }) }
-        var avg: Double { Double(sum)/Double(participants.count) }
+        var sum: Int? {
+            guard points.count == participants.count,
+                  points.count > 0
+            else { return nil }
+            return points.reduce(0, { $0 + $1.value })
+        }
+        var avg: Double? {
+            guard let sum = sum else { return nil }
+            return Double(sum)/Double(participants.count)
+        }
     }
 }
 
@@ -26,8 +34,8 @@ extension UserStory.Vote {
     struct Encoded: Encodable {
         var participants: [String]
         var points: [String: Int]
-        var sum: Int
-        var avg: Double
+        var sum: Int?
+        var avg: Double?
     }
     var encoded: Encoded { Encoded(
         participants: participants,
